Injecting fixtures is much easier in pytest 8.1!
Move logic for older pythons in the `compat.py` module.
