+++ src/pytest_bdd/steps.py
@@ -201,45 +201,3 @@ def find_unique_name(name: str, seen: Iterable[str]) -> str:
         new_name = f"{name}_{i}"
         if new_name not in seen:
             return new_name
-
-
-def inject_fixture(request: FixtureRequest, arg: str, value: Any) -> None:
-    """Inject fixture into pytest fixture request.
-
-    :param request: pytest fixture request
-    :param arg: argument name
-    :param value: argument value
-    """
-
-    fd = compat.FixtureDef(
-        fixturemanager=request._fixturemanager,
-        baseid=None,
-        argname=arg,
-        func=lambda: value,
-        scope="function",
-        params=None,
-    )
-    fd.cached_result = (value, 0, None)
-
-    old_fd = request._fixture_defs.get(arg)
-    add_fixturename = arg not in request.fixturenames
-
-    def fin() -> None:
-        # TODO: Still required?
-        request._fixturemanager._arg2fixturedefs[arg].remove(fd)
-
-        if old_fd is not None:
-            request._fixture_defs[arg] = old_fd
-
-        if add_fixturename:
-            request._pyfuncitem._fixtureinfo.names_closure.remove(arg)
-
-    request.addfinalizer(fin)
-
-    # inject fixture definition
-    request._fixturemanager._arg2fixturedefs.setdefault(arg, []).append(fd)
-
-    # inject fixture value in request cache
-    request._fixture_defs[arg] = fd
-    if add_fixturename:
-        request._pyfuncitem._fixtureinfo.names_closure.append(arg)
